scripts/requirements.txt
aiohttp>=3.8.5
backoff>=2.2.1
requests>=2.31.0
asyncio>=3.4.3
python-dotenv>=1.0.0
sentence-transformers>=2.2.2  # For vector embeddings in memory manager 
pyyaml>=6.0.1
jsonschema>=4.19.1 